move filter_plot_points function
